Complete MockK migration Phase 4: consolidate 17 inline mock factories
Consolidates all inline MockK factory functions from 4 test files into the
central TrackTestMocks.kt repository, completing the Phase 4 factory
consolidation plan.

Changes:
- Added 13 new factory functions to TrackTestMocks.kt
- Resolved 3 duplicate factories (2 merged, 1 split):
  * createMockTrack() - merged with optional maxSpeed parameter
  * createMockPath() - merged returning ArrayPath
  * createMockSemaphore() - split into createMockSemaphoreReal() and
    createMockSemaphoreMock() (incompatible implementations)
- Removed 157 lines of duplicated code from 4 test files:
  * DeadlockDetectionTest.kt - removed 7 factories
  * TrainPathInteractionTest.kt - removed 4 factories
  * DefaultRailWayNetGridTest.kt - removed 1 factory
  * AnimatedSimulationCellRendererTest.kt - removed 3 factories
- Organized into 5 categories:
  * Track Facilities (6 factories)
  * Semaphores (4 factories)
  * Path & Network Elements (3 factories)
  * Occupants & Nodes (2 factories)
  * Mock Implementations (1 class)

Test results: ✅ All 1845 unit tests passing, 307 integration tests passing
Code quality: ✅ detekt and ktlintCheck passing

Outcome: Single centralized mock factory repository with consistent
organization. Zero test regressions. Issue #332 (Phases 1-4) complete.
